[ fix ] 페이지 분리
